Feat/items (#5)
* add todo_list items endpoints

* feat(todo list):
- index of lists
- create a list
- destroy a list
- error handling

* feat(todo list item):
- create an item
- index of all todo list items
- delete an item
