+++ app/controllers/api/v1/items_controller.rb
@@ -1,8 +1,6 @@
 module Api
   module V1
-    class ItemsController < ApplicationController
-      before_action :authenticate_user!
-      before_action :set_current_user
+    class ItemsController < BaseController
 
       def index
         @items = todo_list.items
@@ -10,29 +8,14 @@ module Api
 
       def create
         @item = todo_list.items.build(item_params)
-        @item.save
+        @succesfuly_created = @item.save
+
+      rescue ActiveRecord::RecordNotFound => e
+        @succesfuly_created = false
+        @error = e.message.split(' [')[0]
       end
 
       private
-
-      attr_reader :current_user
-
-      def set_current_user
-        jwt_payload = JWT.decode(
-          request.headers['Authorization'].split(' ').last,
-          ENV['DEVISE_JWT_SECRET_KEY']
-        ).first
-
-        @current_user = User.find(jwt_payload['sub'])
-      end
-
-      def failed_creation
-        render json: {
-          status: 422,
-          message: 'Todo list could not be created.',
-          errors: @todo_list.errors.full_messages.join(', ')
-        }, status: :unprocessable_entity
-      end
 
       def todo_list_params
         params.require(:todo_list).permit(:title)
