+++ app/controllers/api/v1/todo_lists_controller.rb
@@ -1,8 +1,6 @@
 module Api
   module V1
-    class TodoListsController < ApplicationController
-      before_action :authenticate_user!
-      before_action :set_current_user
+    class TodoListsController < BaseController
 
       def index
         @todo_lists = current_user.todo_lists
@@ -10,34 +8,22 @@ module Api
 
       def create
         @todo_list = current_user.todo_lists.build(todo_list_params)
-        return failed_creation unless @todo_list.save
+        @successfully_created = @todo_list.save
       end
 
       def destroy
+        @successfully_deleted = true
         @todo_list = current_user.todo_lists.find(params[:id])
-        @todo_list.destroy
+        return if @todo_list.destroy
+
+        @successfully_deleted = false
+
+      rescue ActiveRecord::RecordNotFound => e
+        @successfully_deleted = false
+        @error = e.message.split(' [')[0]
       end
 
       private
-
-      attr_reader :current_user
-
-      def set_current_user
-        jwt_payload = JWT.decode(
-          request.headers['Authorization'].split(' ').last,
-          ENV['DEVISE_JWT_SECRET_KEY']
-        ).first
-
-        @current_user = User.find(jwt_payload['sub'])
-      end
-
-      def failed_creation
-        render json: {
-          status: 422,
-          message: 'Todo list could not be created.',
-          errors: @todo_list.errors.full_messages.join(', ')
-        }, status: :unprocessable_entity
-      end
 
       def todo_list_params
         params.require(:todo_list).permit(:title)
