+++ app/views/api/v1/todo_lists/create.json.jbuilder
@@ -1,6 +1,22 @@
-json.status :ok
+if @successfully_created
 
-json.todoList do
-  json.id @todo_list.id
-  json.title @todo_list.title
+  json.status do
+    json.code 200
+    json.message 'todo list created successfully.'
+  end
+
+  json.todoList do
+    json.id @todo_list.id
+    json.title @todo_list.title
+  end
+
+else
+
+  json.status do
+    json.code 422
+    json.message 'todo list could not be created.'
+  end
+
+  json.errors @todo_list.errors.full_messages.join(', ')
+
 end
