+++ app/views/api/v1/todo_lists/destroy.json.jbuilder
@@ -1,6 +1,20 @@
-json.status :ok
+if @successfully_deleted
 
-json.todoList do
-  json.id @todo_list.id
-  json.title @todo_list.title
+  json.status do
+    json.code 200
+    json.message 'todo list deleted successfully.'
+  end
+
+else
+
+  json.status do
+    json.code 422
+    json.message 'todo list could not be deleted.'
+  end
+
+  if @error
+    json.error @error
+  else
+    json.errors @todo_list.errors.full_messages.join(', ')
+  end
 end
