+++ app/views/api/v1/todo_lists/index.json.jbuilder
@@ -1,4 +1,7 @@
-json.status :ok
+json.status do
+  json.code 200
+  json.message 'todo lists retrieved successfully.'
+end
 
 json.todoLists @todo_lists do |todo_list|
   json.id todo_list.id
