+++ config/routes.rb
@@ -11,8 +11,8 @@ Rails.application.routes.draw do
 
   namespace :api do
     namespace :v1 do
-      resources :todo_lists, only: %i[index create update destroy] do
-        resources :items, only: %i[index create update destroy]
+      resources :todo_lists, only: %i[index create destroy] do
+        resources :items, only: %i[index create destroy]
       end
     end
   end
